+++ spring-cloud-stream-binder-sqs/src/main/java/org/springframework/cloud/stream/sqs/SnsPayloadConvertingChannelInterceptor.java
@@ -1,0 +1,17 @@
+package org.springframework.cloud.stream.sqs;
+
+import org.springframework.messaging.Message;
+import org.springframework.messaging.MessageChannel;
+import org.springframework.messaging.support.ChannelInterceptor;
+import org.springframework.messaging.support.MessageBuilder;
+
+/**
+ * @author Maciej Walkowiak
+ */
+class SnsPayloadConvertingChannelInterceptor implements ChannelInterceptor {
+
+    @Override
+    public Message<?> preSend(Message<?> message, MessageChannel channel) {
+        return MessageBuilder.createMessage(new String((byte[]) message.getPayload()), message.getHeaders());
+    }
+}
